server/chat/requirements/base.txt
Django>=2.2.0,<2.3.0
djangorestframework>=3.6.2,<4
